Fix issue with NuCache knowing the web project
