+++ src/Umbraco.PublishedCache.NuCache/Umbraco.PublishedCache.NuCache.csproj
@@ -13,7 +13,6 @@
     <ItemGroup>
         <ProjectReference Include="..\Umbraco.Abstractions\Umbraco.Abstractions.csproj" />
         <ProjectReference Include="..\Umbraco.Infrastructure\Umbraco.Infrastructure.csproj" />
-        <ProjectReference Include="..\Umbraco.Web\Umbraco.Web.csproj" />
     </ItemGroup>
 
     <ItemGroup>
